requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
faiss-cpu
pypdf
python-dotenv
mcp
